requirements.txt
# requirements.txt

openai>=1.0.0       # Use a recent version
numpy>=1.21.0
matplotlib>=3.4.0
pandas>=1.3.0
seaborn>=0.11.0
scipy>=1.7.0
rich>=10.0.0
python-dotenv>=0.19.0 # Add this line
web3>=6.0.0          # If using blockchain features